+++ code/ryzom/server/src/ai_share/ai_actions.h
@@ -267,7 +267,7 @@ public:
 						if (index>=data.size())
 						{
 							nlwarning("Unexpected end of input data in serialFromString()");
-							index=data.size()+1;
+							index=(uint)data.size()+1;
 							return CArg();
 						}
 					}
@@ -279,7 +279,7 @@ public:
 					if (index>data.size())
 					{
 						nlwarning("Unexpected end of input data in serialFromString()");
-						index=data.size()+1;
+						index=(uint)data.size()+1;
 						return CArg();
 					}
 					return CArg(bool(data[index] == 1));
@@ -289,7 +289,7 @@ public:
 					if (index>data.size())
 					{
 						nlwarning("Unexpected end of input data in serialFromString()");
-						index=data.size()+1;
+						index=(uint)data.size()+1;
 						return CArg();
 					}
 					return CArg(((sint32 *)&data[index])[-1]);
@@ -299,7 +299,7 @@ public:
 					if (index>data.size())
 					{
 						nlwarning("Unexpected end of input data in serialFromString()");
-						index=data.size()+1;
+						index=(uint)data.size()+1;
 						return CArg();
 					}
 					return CArg(((float *)&data[index])[-1]);
@@ -315,7 +315,7 @@ public:
 					break;
 			}
 			nlwarning("Unexpected type in serialFromString()");
-			index=data.size()+1;
+			index=(uint)data.size()+1;
 			return CArg();
 		}
